app/requirements.txt
alembic==1.7.7
anyio==3.5.0
asgiref==3.5.0
click==8.1.3
colorama==0.4.4
fastapi==0.76.0
FastAPI-SQLAlchemy==0.2.1
greenlet==1.1.2
h11==0.13.0
idna==3.3
importlib-metadata==4.11.3
importlib-resources==5.7.1
joblib==1.1.0
Mako==1.2.0
MarkupSafe==2.1.1
numpy==1.22.3
pandas==1.1.5
pip==22.0.4
psycopg2==2.9.3
pydantic==1.9.0
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.1
scikit-learn==0.24.2
setuptools==59.8.0
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.4.36
starlette==0.18.0
typing_extensions==4.2.0
uvicorn==0.16.0
wheel==0.37.1
zipp==3.8.0